ByoYomiLabel Change
It is not a norm of any go server to have a byo yomi label. Many players find it distracting, most notably me. https://forums.online-go.com/t/option-to-remove-countdown-inside-stone/28209/6

The most elegant solution would be to remove it in favor of a flashing clock, which is 99% implemented here with a small stylesheet issue causing no flash to display.

A creative commons zero icon created by my is also included should you want to spend the effort to make this a user preference instead of change the sitewide behavior permanently. Thank you